Support creating users on submit_as from backend
When users only existed in the authentication backend and they were
specified as submit_as in post_review, the web api simply returned a no
user error. Now, instead we ask the backend for a user if there is none
in the database and everything is created as we need.